From 5ae9ddb9a4086c63a50bd3c4c96a06fad04f3be9 Mon Sep 17 00:00:00 2001 [PATCH] x86/HVM: Avoid cache flush operations during hvm_load
An MTRR record is processed for each vCPU during hvm_load. Each MTRR
record sets several mtrrs, each of which flushes the cache on all pCPUs.
This can take some time and trip the watchdog for HVM guests with many
CPUs.

To fix this, introduce a flag which prevents flushing the cache when
doing MTRR operations and instead do a flush at the end of hvm_load.

This reduces the time to restore an HVM guest with 32 vCPUs by about 5
seconds.

Signed-off-by: Ross Lagerwall <ross.lagerwall@citrix.com>